add debug for plugin and ioredis, rename stack
