+++ server/bootstrap.js
@@ -1,6 +1,18 @@
+const debug = require('debug')
+
 module.exports = async ({ strapi }) => {
   // Load plugin Config
   const coreConfig = strapi.config.get('plugin.redis');
+
+  // Configure plugin debug
+  if (coreConfig.settings.debug === true) {
+    debug.enable('strapi:strapi-plugin-redis')
+  }
+
+  // Allow plugin + ioredis debug
+  if (coreConfig.settings.debug === true && coreConfig.settings.debugIORedis === true) {
+    debug.enable('strapi:strapi-plugin-redis,ioredis:*')
+  }
 
   // Construct Redis API
   strapi.redis = {
